refactor: remove circular dependencies from `ui/components/*`
