+++ development/circular-deps.jsonc
@@ -17,38 +17,6 @@
     "ui/components/app/name/name.tsx"
   ],
   [
-    "ui/components/multichain/pages/send/components/account-picker.tsx",
-    "ui/components/multichain/pages/send/components/index.ts"
-  ],
-  [
-    "ui/components/multichain/pages/send/components/address-book.tsx",
-    "ui/components/multichain/pages/send/components/index.ts"
-  ],
-  [
-    "ui/components/multichain/pages/send/components/hex.tsx",
-    "ui/components/multichain/pages/send/components/index.ts"
-  ],
-  [
-    "ui/components/multichain/pages/send/components/index.ts",
-    "ui/components/multichain/pages/send/components/network-picker.tsx"
-  ],
-  [
-    "ui/components/multichain/pages/send/components/index.ts",
-    "ui/components/multichain/pages/send/components/recipient-content.tsx"
-  ],
-  [
-    "ui/components/multichain/pages/send/components/index.ts",
-    "ui/components/multichain/pages/send/components/recipient-input.tsx"
-  ],
-  [
-    "ui/components/multichain/pages/send/components/index.ts",
-    "ui/components/multichain/pages/send/components/recipient.tsx"
-  ],
-  [
-    "ui/components/multichain/pages/send/components/index.ts",
-    "ui/components/multichain/pages/send/components/your-accounts.tsx"
-  ],
-  [
     "ui/pages/notifications/notifications-list.tsx",
     "ui/pages/notifications/notifications.tsx"
   ]
