+++ ui/components/multichain/pages/send/components/address-book.tsx
@@ -26,7 +26,7 @@ import {
   MetaMetricsEventCategory,
   MetaMetricsEventName,
 } from '../../../../../../shared/constants/metametrics';
-import { SendPageRow } from '.';
+import { SendPageRow } from './send-page-row';
 
 export const SendPageAddressBook = () => {
   const t = useContext(I18nContext);
